#3529 Captions/depictions are not saved to Commons - remove captions interface
