+++ app/src/main/java/fr/free/nrw/commons/upload/WikiBaseInterface.java
@@ -1,19 +1,17 @@
 package fr.free.nrw.commons.upload;
 
-import androidx.annotation.NonNull;
+import static org.wikipedia.dataclient.Service.MW_API_PREFIX;
 
+import androidx.annotation.NonNull;
+import io.reactivex.Observable;
 import org.wikipedia.dataclient.mwapi.MwPostResponse;
 import org.wikipedia.dataclient.mwapi.MwQueryResponse;
-
-import io.reactivex.Observable;
 import retrofit2.http.Field;
 import retrofit2.http.FormUrlEncoded;
 import retrofit2.http.GET;
 import retrofit2.http.Headers;
 import retrofit2.http.POST;
 import retrofit2.http.Query;
-
-import static org.wikipedia.dataclient.Service.MW_API_PREFIX;
 
 /**
  * Retrofit calls for managing responses network calls of entity ids required for uploading depictions
@@ -30,5 +28,19 @@ public interface WikiBaseInterface {
 
     @GET(MW_API_PREFIX + "action=query&prop=info")
     Observable<MwQueryResponse> getFileEntityId(@Query("titles") String fileName);
+
+    /**
+     * Upload Captions for the image when upload is successful
+     *
+     * @param FileEntityId enityId for the uploaded file
+     * @param editToken editToken for the file
+     * @param captionValue value of the caption to be uploaded
+     */
+    @FormUrlEncoded
+    @POST(MW_API_PREFIX + "action=wbsetlabel&language=en")
+    Observable<MwPostResponse> addLabelstoWikidata(@Field("id") String FileEntityId,
+        @Field("token") String editToken,
+        @Field("language") String language,
+        @Field("value") String captionValue);
 
 }
